fix(v2.0.48): coingecko INSERT never persisted — NULLIF($3, 0) inferred the decimal param as INTEGER
pgx sends shopspring decimals through driver.Valuer as text; inside
NULLIF($3, 0) the untyped 0 literal makes Postgres infer $3 as integer,
so every snapshot since v2.0.44 failed with SQLSTATE 22P02 (seen in prod
logs at 2.0.47 startup: 'invalid input syntax for type integer:
59.141936') — coingecko_snapshots stayed empty and the macro gate /
dominance slope ran fail-open the whole time. Explicit ::numeric casts
on all five params. Reproduced and verified against postgres 16
(NULLIF('59.141936',0) errors; NULLIF('59.141936'::numeric,0) works).

+++ backend/internal/marketdata/coingecko.go
@@ -94,7 +94,7 @@ func (c *Collector) collectCoinGecko(ctx context.Context) {
 	if _, err := c.db.Exec(ctx, `
 		INSERT INTO coingecko_snapshots
 			(btc_usd, btc_24h_pct, btc_dominance_pct, total_mcap_usd, mcap_24h_pct)
-		VALUES ($1, $2, NULLIF($3, 0), NULLIF($4, 0), $5)
+		VALUES ($1::numeric, $2::numeric, NULLIF($3::numeric, 0), NULLIF($4::numeric, 0), $5::numeric)
 	`,
 		decimal.NewFromFloat(snap.BTCUSD).Round(4),
 		decimal.NewFromFloat(snap.BTC24hPct).Round(6),
